Rename beatertest to monitoringtest
